pyproject.toml

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "numba_scfg"
version = "0.0.6-dev0"
authors = [
    {name = "Numba Developers", email="none@none.none"},
]
description = "Numba Compatible SCFG utilities"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["numba"]
license = {text = "BSD-2-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "graphviz",
    "pyyaml",
]

[tool.setuptools.packages]
find = {}

[project.optional-dependencies]
dev = [
    "sphinx",
    "myst-parser",
    "furo",
    "pytest",
    "sphinx-rtd-theme"
]
